pyproject.toml
[project]
name = "ai-code-review-api"
version = "0.1.0"
description = "FastAPI service for AI-powered code review."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0",
    "openai>=1.35.0",
    "anthropic>=0.30.0",
    "google-generativeai>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
]
